requirements.txt
asgiref==3.7.2
attrs==23.1.0
coverage==7.3.2
Django==4.2.5
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Markdown==3.4.4
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
